Move files around in preparation for refactor
