+++ src/builder/fast-path.js
