+++ src/common/deprecated.js
