+++ src/common/parser-include-shebang.js
