+++ src/config/resolve-config-editorconfig.js
