+++ src/debug/clean-ast.js
