+++ src/language-css/parser-postcss.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const createError = require("./parser-create-error");
+const createError = require("../common/parser-create-error");
 
 function parseSelector(selector) {
   const selectorParser = require("postcss-selector-parser");
